refactor: Place Router to controller dir
